Simplify RNG
It was complicated because I wanted to avoid using hardware RNG
as much as possible. But then, I discovered that you cannot
generate EC keys with your custom RNG, so there isn't a point in
using this RNG design.

Still, I leave the class in case I want to bring it back.